+++ sqlbuilder_insert.go
@@ -61,18 +61,21 @@ func (ib *InsertBuilder) SetMap(m map[string]any, opts ...BuilderOption) *Insert
 		return ib
 	}
 
-	bo := &BuilderOptions{}
+	bo := &BuilderOptions{
+		DbColumns: make(map[string]bool),
+	}
 	for _, opt := range opts {
 		opt(bo)
 	}
 
 	for n, v := range m {
-		if bo.ToSnake == nil {
-			ib.Set(n, v)
-		} else {
-			ib.Set(bo.ToSnake(n), v)
+		if bo.ToSnake != nil {
+			n = bo.ToSnake(n)
 		}
 
+		if _, ok := bo.DbColumns[n]; ok {
+			ib.Set(n, v)
+		}
 	}
 	return ib
 }